requirements.txt
pycryptodome
pdf417gen
Pillow
pdf417decoder
